Added documentation for bootloader
The bootloader was sorely lacking documentation as the only records of how
memory is supposed to be laid out were strewn across various pieces of paper
under my desk.

I realize the bootloader doesn't 100% follow what's laid out in
the docs, and many of these items are marked "Not Implemented" or otherwise
noted. If there is a discrepancy between what the bootloader does and the
documentation, the documentation is correct.

I also removed the "Better Documentation" entry in `todo.txt`. I realize the
documentation still isn't great, but my goal is to make good documentation an
ongoing thing and don't want to think about it as ever being done.