Completed keymap and added functions to use to
The keymap has been filled out and documented. The keymap is used to convert the
responses from a PS2 keyboard device to the format used by AndromedOS. Functions
to convert to ASCII and wait for a key to be pressed have been added. The
terminal_boxread function still needs work.