Stop confusing mouse and keyboard input
If bit 5 of the status register is set the input is assumed to be coming from
the mouse.

This only seems to work when the status register is accessed after the response
register. Crude testing shows this seems to work, but it's possible mouse input
may still be interpreted as keyboard ( or vice versa ) in certain situations.